Add files via upload
무기 획득 시 무기마다 다른 총알 수를 가지도록 만듦
- 총알도 5.56, 7.62, 12게이지 별로 다른 데미지로 변경할 것임

+++ WeaponBase.h
@@ -49,5 +49,5 @@ public:
 	UPROPERTY(VisibleAnywhere, BlueprintReadOnly, Category = "Weapon Info: Name")
 		FName Weapon_Name; // 무기 이름
 	UPROPERTY(VisibleAnywhere, BlueprintReadOnly, Category = "Weapon Info: Ammo")
-		float W_Ammo; // 무기 총알
+		int32 W_Ammo; // 무기 총알
 };
